+++ ML-Frameworks/pytorch-aarch64/versions.sh
@@ -9,15 +9,15 @@
 # For information on how to update the versions below, read the README.md.
 
 # get-source.sh deps
-PYTORCH_HASH=004acb4e35805062a643ee1588cdae584b2b2957   # 2.13.0.dev20260504 from viable/strict, May 4th, 2026
-IDEEP_HASH=6469a610baaa94ba15de1902fd1afb25316171d2     # From ideep_pytorch, Apr 23rd, 2026
-ONEDNN_HASH=8fa20dc93a75c094f618f3f8c206775731d63301    # From main, May 5th, 2026
-KLEIDIAI_HASH=7a7da265d21f42053d453ef664814c5b5cad8cd3  # v1.24.0 from main, May 4th, 2026
+PYTORCH_HASH=c230e6986c6aaee101e24ca14813c21add0b148f   # 2.13.0.dev20260528 from viable/strict, May 28th, 2026
+IDEEP_HASH=e087b6e4b32a7ba684db82231d1558123968ac1d     # From ideep_pytorch, May 11th, 2026
+ONEDNN_HASH=3004f0a1d9cf92c06eaaca57840aaa2149ebba85    # From main, May 27th, 2026
+KLEIDIAI_HASH=5866364d3bc079d2d6cae5f0acf6d076594bc7a7  # v1.25.0 from main, May 28th, 2026
 
 # build-wheel.sh deps
-ACL_VERSION="v53.0.0"   # Apr 10th
+ACL_VERSION="v53.1.0"   # May 18th
 OPENBLAS_VERSION="v0.3.33"  # Apr 23rd
 
 # Dockerfile deps
-TORCHVISION_NIGHTLY="0.27.0.dev20260505"
-TORCHAO_NIGHTLY="0.18.0.dev20260505"
+TORCHVISION_NIGHTLY="0.28.0.dev20260527"
+TORCHAO_NIGHTLY="0.18.0.dev20260528"